Start libvirt lazy to allow setting up the pod
In case that virt-handler sets up devices or other things directly
inside the container, it is important that libvirt is started after all
that setup is done. Otherwise libvirt may detect not all provided
virtualization capabilities inside the node.

Libvirt will be started when virt-handler tries to create a VMI. On the
first try, virt-launcher will block a little bit longer to give libvirt
the chance to start. If that times out, subsequent reconnect attempts
are much shorter, in order to not block virt-handler workers for too
long.